requirements.txt
PyQt5
Pillow
tensorflow==2.7
opencv-python
tensorflow-addons
sklearn
matplotlib
qdarkstyle
pyinstaller